server/requirements.txt
anyio==3.5.0
asgiref==3.5.0
bcrypt==3.2.0
black==22.1.0
cffi==1.15.0
click==8.0.3
cryptography==36.0.1
fastapi==0.73.0
greenlet==1.1.2
h11==0.13.0
httptools==0.3.0
idna==3.3
mypy==0.931
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.0
psycopg2==2.9.3
pycparser==2.21
pydantic==1.9.0
PyJWT==2.3.0
python-dotenv==0.19.2
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.31
sqlalchemy2-stubs==0.0.2a19
sqlmodel==0.0.6
starlette==0.17.1
tomli==2.0.1
typing_extensions==4.0.1
uvicorn==0.17.4
uvloop==0.16.0
watchgod==0.7
websockets==10.1
